fix branch behavior: use existing branches.
we used to treat creating a new branch and switching
to an existing branch as the same, using the current
base commit. this is wrong.

now, if the branch we're switching to exists and we
have no '-b' parameter, we use its current commit as
the commit we switch to.