speed up git/walk massively
Don't stat the files on git/fs from git/walk: this
is not very fast, since on the first access (and if
we fall out of cache), we need to decompress all
the data. This is slow.

Only looking at the qid takes git/walk from 20+
seconds on the perl repository to just about 1
second.