packfiles: make cache size tunable.
For serving, the cache is mostly useless, since we
do a walk of all objects, ideally once; we also tend
to hog a bunch of ram and possibly fall over if there
are a bunch of concurrent clones.

To make serving better, allow tuning the cache size
via an environment variable.