import, export: fix patch formatting
Git uses '---' to separate the message from the patch.
We used to, but we were buggy about it -- and then a
recent patch broke it entirely.

This reverts and fixes the previous behavior, so that

	git/export > /tmp/patch
	git/branch -nb HEAD~ master
	git/import /tmp/patch

will round trip properly, and produce the same commit
id. This makes it easier to send around patches, since
applying a patch will not cause conflicts. It should
also make for a nicer git/rebase.