git/import: avoid adding nonexistent files
When importing patches with hunk offsets, some extra lines
are printed when files are patched. Since we use the output
to know which files to add or remove, this can lead to an
attempt to add nonexistent files.

Filter down to just the output about patching, so we don't
print spurious messages.