git/rebase: first cut of rebase implementation
Implement the basic version of git/rebase, handling
rewriting commits on one branch as though they were
based off a different branch.

Rebasing starts by finding the least common ancestor
of the two branches, and then replaying every commit
from the source branch on to the destination branch,
then renaming the source branch.

In the case of a conflict, the commits are dumped to
a temp file, and the rebase is paused until the user
either aborts the rebase or resolves the issue and
resumes the rebase.