git/send: smaller packfiles when pushing
When pushing to a remote repository, we would
pack at least all objects in one commit, even
if the remote repository had most of them.

This change prunes out all objects from the
last commit that they have, which should drop
the size of the generated pack file noticably
for the common case of pushing a commit or
that changes a small percentage of files
in a larger repository.