git/branch: implcitly create local branch off origin
Currently, creating a local branch off a remote one requires
passing the '-nb origin/$branch' argument to git/branch.

In addition to being annoying, it opens up the possibility for
human error, since passing this argument for an existing branch
will clobber it.

After this change, a repo in this state:

	% git/branch -a
	heads/master
	remote/origin/master
	remote/origin/foo

will switch to heads/master when running:

	% git/branch master

but will create a new branch foo, based off origin/foo,
when running:

	% git/branch foo