packfiles: optimize access to objects
We were still doing a lot of spare seeking and file
access. This removes 2 more major sources of it: we
no longer open the pack directory to list it unless
there's a lookup failure, and we try the packfiles
before we try the loose object, saving pointless file
opens and seeks.