git/init: change default branch, remove template
The stuff in the template directory is mostly
unused by git9, so we may as well get rid of it.

While we're here, remove support for creating
bare repos, since none of the rest of our stuff
supports it, and having an extra '.git/' in an
empty dir isn't worth the trouble.

Also while we're here, change the default branch
to 'front': it's more playful than 'main', and
avoids debates about the default name.