git/serve: set HEAD on first push
If there are no valid refs, and HEAD is invalid, then
pick the ref with the newest commits as the default
branch.

Several people have been caught out by pushing to
a repo where HEAD named differently from what got
pushed, and this is going to be more of a footgun
when 'master', 'main', and 'front' are all in active
use. This should make us pick a useful default in
those cases, instead of silently failing.