git/query: exclude left endpoint from range
This is closer to git behavior and is more flexible since the original
behavior can be recovered with a~..b, but not the other way around.

git/rebase needs these semantics since we do not want to include the
common ancestor in the list of commits to import onto the new base.
The new base may be newer than the oldest commit in the import list,
so it can appear anywhere in the current git/query output and cannot
be pruned by git/rebase itself.