git/merge, git/branch: share merge code, delete better
There were some edge cases when merges deleted files:
With diff3 redirected to an output, the output always
exists. This lead to trash files being added to repos,
and branch switches not cleaning up properly.

In addition, this speeds up git/merge by not walking
the whole tree, just the files changed by the commits
that are getting merged.