git/clone: allow cloning a single branch.
Sometimes you only need one branch from a repo. This
was a pain in the ass before, involving initializing
the repository, pulling to get the hashes, and then
manually creating the branch because the empty repo
had no valid branches, so branch switchnig broke.

The change to support this was trivial, so why not.